+++ js/comment-max-symbol.js
@@ -1,0 +1,10 @@
+// проверка максимальной длины строки
+const maxLengthComment = 140;
+
+const stringLength = (string) => {
+  if (typeof string === 'string') {
+    return string.length <= maxLengthComment;
+  }
+};
+
+export {stringLength};
